update CR comments and move logic ourside of kafka consumer to Kafka event listener
